Lennard-Jones Force Fields and Potential Energy

Lennard-Jones force field parameters are stored in comma-separated-value format in PorousMaterials.PATH_TO_FORCEFIELDS.

Interaction of an adsorbate with the crystal is modeled as pair-wise additive and with Lennard-Jones potentials of the form:

V(r) = 4 * ϵ * [ x ^ 12 - x ^ 6 ], where x = σ / r

The Lennard-Jones force field input files, e.g. UFF.csv contain a list of pure (i.e. X-X, where X is an atom) sigmas (σ) and epsilons (ϵ) with units Angstrom (Å) and Kelvin (K), respectively. Note that, e.g., in the UFF paper, the Lennard-Jones potential is written in a different form; thus, parameters need to be converted to correspond to the functional form used in PorousMaterials.jl.

Building Blocks of PorousMaterials: Lennard-Jones Force Fields

Loading Force Field Files and Accessing Attributes

Reading in Lennard-Jones force field parameters is made easy with the LJForceField function. Let's load in the parameters from the Universal Force Field file (UFF.csv):

using PorousMaterials

# read in Lennard-Jones force field parameters from the Universal Force Field
ljforcefield = LJForceField("UFF", r_cutoff=14.0, mixing_rules="Lorentz-Berthelot")

PorousMaterials will then output information about the force field file you just loaded:

    Force field: UFF
    Number of atoms included: 106
    Cut-off radius (Å) = 14.0

This also prints all of the atoms included in the loaded forcefield with their given ϵ and σ. This was excluded because it would use too much space on this page.

We can access attributes LJForceField such as pure_σ, pure_ϵ, and interaction values:

# access the Lennard-Jones epsilon & sigma for Xe 
ljforcefield.pure_ϵ[:Xe] # K
ljforcefield.pure_σ[:Xe] # Å

# access the Lennard-Jones epsilon & sigma for Xe-C interactions
ljforcefield.ϵ[:Xe][:C]  # K
ljforcefield.σ²[:Xe][:C] # Å (store σ² for faster computation)

Checking Force Field Coverage

When running simulations, it is necessary to have the force field terms for all of the atoms. This can be checked using forcefield_coverage:

using PorousMaterials

ljforcefield = LJForceField("UFF")

xtal = Crystal("SBMOF-1.cif")
strip_numbers_from_atom_labels!(xtal)

# check is the atoms in the crystal are covered
forcefield_coverage(xtal, ljforcefield)

molecule = Molecule("CO2")

# check if the atoms in the molecule are covered
forcefield_coverage(molecule, ljforcefield)

Simulation Box and the Cutoff Radius

Find the replication factors needed to make a supercell big enough to fit a sphere with the specified cutoff radius.In PorousMaterials.jl, rather than replicating the atoms in the home unit cell to build the supercell that serves as a simulation box, we replicate the home unit cell to form the supercell (simulation box) in a for loop.The replication_factors function ensures enough replication factors such that the nearest image convention can be applied.

using PorousMaterials

ljforcefield = LJForceField("UFF")

xtal = Crystal("SBMOF-1.cif")

r_cutoff = 14.0 # Å
repfactors = replication_factors(xtal.box, r_cutoff) 

Potential Energies: Van der Waals

What is the van der Waals potential energy of a Xe adsorbate inside SBMOF-1 at Cartesian coordinates [0.0, 1.0, 3.0] using the UFF as a molecular model?

using PorousMaterials

xtal = Crystal("SBMOF-1.cif")
strip_numbers_from_atom_labels!(xtal)

ljforcefield = LJForceField("UFF")

# load molecule and convert it to fractional
molecule = Molecule("Xe")
molecule = Frac(molecule, xtal.box) 

translate_to!(molecule, Cart([0.0, 1.0, 3.0]), xtal.box) # need box b/c we're in Cartesian

energy = vdw_energy(xtal, molecule, ljforcefield) # K

detailed docs

Forcefields

PorousMaterials.LJForceField — Type

Data structure for a Lennard Jones forcefield.

Attributes

name::String: name of forcefield; correponds to filename

pure_σ::Dict{Symbol, Float64}: Dictionary that returns Lennard-Jones σ of an X-X interaction, where X is an atom. (units: Angstrom)

pure_ϵ::Dict{Symbol, Float64}: Dictionary that returns Lennard-Jones ϵ of an X-X interaction, where X is an atom. (units: K)

σ²::Dict{Symbol, Dict{Symbol, Float64}}: Lennard Jones σ² (units: Angstrom²) for cross-interactions. Example use is sigmas_squared[:He][:C]

ϵ::Dict{Symbol, Dict{Symbol, Float64}}: Lennard Jones ϵ (units: K) for cross-interactions. Example use is epsilons[:He][:C]

r²_cutoff::Float64: The square of the cut-off radius beyond which we define the potential energy to be zero (units: Angstrom²). We store σ² to speed up computations, which involve σ², not σ.

PorousMaterials.replication_factors — Function

repfactors = replication_factors(unitcell, r_cutoff)

Find the replication factors needed to make a supercell big enough to fit a sphere with the specified cutoff radius. In PorousMaterials.jl, rather than replicating the atoms in the home unit cell to build the supercell that serves as a simulation box, we replicate the home unit cell to form the supercell (simulation box) in a for loop. This function ensures enough replication factors such that the nearest image convention can be applied.

A non-replicated supercell has 1 as the replication factor in each dimension (repfactors = (1, 1, 1)).

Arguments

unitcell::Box: The unit cell of the crystal

r_cutoff::Float64: Cutoff radius beyond which we define the potential energy to be zero (units: Angstrom)

Returns

repfactors::Tuple{Int, Int, Int}: The replication factors in the a, b, c directions

PorousMaterials.forcefield_coverage — Function

forcefield_coverage(atoms, ljforcefield)
forcefield_coverage(molecule, ljforcefield)
forcefield_coverage(crystal, ljforcefield)

Check that the force field contains parameters for every species in atoms::Atoms. Will print out which atoms are missing.

Arguments

atoms::Atoms: a set of atoms

ljforcefield::LJForceField: A Lennard Jones forcefield object containing information on atom interactions

Returns

all_covered::Bool: returns true if all species in the atoms are covered by the force field.

Potential Energy

Missing docstring.

Missing docstring for PotentialEnergy. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SystemPotentialEnergy. Check Documenter's build log for details.

Nearest Image Conventions

PorousMaterials.nearest_image! — Function

nearest_image!(dxf)

Applies the nearest image convention on a vector dxf between two atoms in fractional space; modifies dxf for nearest image convention. Fractional coordinates here fall in [0, 1] so that the box is [0, 1]^3 in fractional space.

Warning: this assumes the two molecules are in the box described by fractional coords [0, 1]³.

Arguments

dxf::Array{Float64}: A vector between two atoms in fractional space

Electrostatics Energy

PorousMaterials.Eikr — Type

eikr = Eikr(eikar, eikbr, eikcr)

mutable struct for holding the eikr vectors

Attributes

eikar::OffsetArray{Complex{Float64}}: array for storing e^{i * ka ⋅ r}; has indices 0:kreps[1] and corresponds to recip. vectors in a-direction

eikbr::OffsetArray{Complex{Float64}}: array for storing e^{i * kb ⋅ r}; has indices -kreps[2]:kreps[2] and corresponds to recip. vectors in b-direction

eikcr::OffsetArray{Complex{Float64}}: array for storing e^{i * kc ⋅ r}; has indices -kreps[2]:kreps[1] and corresponds to recip. vectors in c-direction

Missing docstring.

Missing docstring for total. Check Documenter's build log for details.

Missing docstring.

Missing docstring for electrostatic_potential. Check Documenter's build log for details.

PorousMaterials.electrostatic_potential_energy — Function

ϕ = electrostatic_potential_energy(crystal, molecule, eparams, eikr)

Compute the electrostatic potential energy of a molecule inside a crystal.

The electrostatic potential is created by the point charges assigned to the crystal atoms in crystal.charges. Periodic boundary conditions are applied through the Ewald summation. The spurious self-interaction term is neglected here because we are looking at differences in energy in a Monte Carlo simulation.

Warning: it is assumed that the crystal is replicated enough such that the nearest image convention can be applied for the short-range cutoff radius supplied in eparams.sr_cutoff_r.

Arguments

crystal::Crystal: Crystal structure (see crystal.charges for charges)

molecule::Molecule: The molecule being compared to the atoms in the crystal.

eparams::EwaldParams: data structure containing Ewald summation settings

eikr::Eikr: Stores the eikar, eikbr, and eikcr OffsetArrays used in this calculation.

Returns

pot::EwaldSum: Electrostatic potential between crystal and molecule (units: K)

ϕ = electrostatic_potential_energy(molecules, eparams, box, eikr)

Compute the electrostatic potential energy of a system comprised of an array of Molecules.

The EWald summation is used here in a double for loop; do not use this function for Monte Carlo simulations because it is computationally expensive.

Returns an EwaldSum type containing short-range and long-range contributions to the Ewald sum as well as the spurious self-interaction and intramolecular interactions. Access via (ϕ.sr, ϕ.lr, ϕ.self, ϕ.intra).

Units of energy: Kelvin

Arguments

molecules::Array{Molecules, 1}: array of molecules comprising the system.

eparams::EwaldParams: data structure containing Ewald summation settings

box::Box: the box the energy is being computed in

eikr::Eikr: Stores the eikar, eikbr, and eikcr OffsetArrays used in this calculation.

Returns

ϕ::GGEwaldSum: The total electrostatic potential energy

PorousMaterials.precompute_kvec_wts — Function

kvectors = precompute_kvec_wts(kreps, box, α, max_mag_k_sqrd=Inf)

For speed, pre-compute the weights for each reciprocal lattice vector for the Ewald sum in Fourier space. This function takes advantage of the symmetry: cos(-k⋅(x-xᵢ)) + cos(k⋅(x-xᵢ)) = 2 cos(k⋅(x-xᵢ))

If max_mag_k_sqrd is passed, k-vectors with a magnitude greater than max_mag_k_sqrd are not included.

Arguments

kreps::Tuple{Int, Int, Int}: number of k-vector replications required in a, b, c

box::Box: the simulation box containing the reciprocal lattice.

α::Float64: Ewald sum convergence parameter (units: inverse Å)

max_mag_k_sqrd::Float64: cutoff for |k|² in Fourier sum; if passed, do not include

k-vectors with magnitude squared greater than this.

Returns

kvectors::Array{Kvector, 1}: array of k-vectors to include in the Fourier sum and their

corresponding weights indicating the contribution to the Fourier sum.

PorousMaterials.setup_Ewald_sum — Function

eparams = setup_Ewald_sum(box, sr_cutoff_r; ϵ=1e-6, verbose=false)

Given the short-range cutoff radius and simulation box, automatically compute Ewald convergence parameter and number of k-vector replications in Fourier space required for a given precision. Constructs and returns Ewald parameters data type with this information.

Also, pre-compute weights on k-vector contributions to Ewald sum in Fourier space.

Also, allocate OffsetArrays for storing e^{i * k ⋅ r} where r = x - xⱼ and k is a reciprocal lattice vector.

Arguments

box::Box: the simulation box containing the reciprocal lattice.

sr_cutoff_r::Float64: cutoff-radius (units: Å) for short-range contributions to Ewald

ϵ::Float64: desired level of precision. Typical value is 1e-6, but this does not

verbose::Bool: If true will print results

Returns

eparams::EwaldParams: data structure containing Ewald summation settings

corresponding weights indicating the contribution to the Fourier sum.

PorousMaterials.total_electrostatic_potential_energy — Function

total_ϕ = total_electrostatic_potential_energy(molecules, eparams, box, eikr)

Calculates the total electrostatic potential energy of an array of Molecules using a Grand Canonical Monte Carlo (GCMC) algorithm. #TODO add to this

Arguments

molecules::Array{Molecule, 1}: The molecules comprising the system.

eparams::EwaldParams: data structure containing Ewald summation settings

box::Box: The box the energy is being computed in.

eikr::Eikr: Stores the eikar, eikbr, and eikcr OffsetArrays used in this calculation.

Returns

ϕ::GGEwaldSum: The total electrostatic potential energy

total_ϕ = total_electrostatic_potential_energy(crystal, molecules, eparams, eikr)

Explanation of totalelectrostaticpotential_energy that uses crystal

Arguments

crystal::Crystal: Crystal structure (see crystal.charges for charges)

molecules::Array{Molecule, 1}: The molecules comprising the system.

eparams::EwaldParams: data structure containing Ewald summation settings

eikr::Eikr: Stores the eikar, eikbr, and eikcr OffsetArrays used in this calculation.

Van der Waals Energy

PorousMaterials.lennard_jones — Function

energy = lennard_jones(r², σ², ϵ)  (units: Kelvin)

Calculate the lennard jones potential energy given the square of the radius r between two lennard-jones spheres. σ and ϵ are specific to interaction between two elements. Return the potential energy in units Kelvin (well, whatever the units of ϵ are).

Arguments

r²::Float64: distance between two (pseudo)atoms in question squared (Angstrom²)

σ²::Float64: sigma parameter in Lennard Jones potential squared (units: Angstrom²)

ϵ::Float64: epsilon parameter in Lennard Jones potential (units: Kelvin)

Returns

energy::Float64: Lennard Jones potential energy

PorousMaterials.vdw_energy — Function

energy = vdw_energy(crystal, molecule, ljforcefield)

Calculates the van der Waals interaction energy between a molecule and a crystal. Applies the nearest image convention to find the closest replicate of a specific atom.

WARNING: it is assumed that the framework is replicated sufficiently such that the nearest image convention can be applied. See replicate and replication_factors.

ggenergy = vdwenergy(moleculeid, molecules, ljforcefield, simulationbox)

Calculates van der Waals interaction energy of a single adsorbate molecules[molecule_id] with all of the other molecules in the system. Periodic boundary conditions are applied, using the nearest image convention.

Arguments

molecule_id::Int: Molecule ID used to determine which molecule in molecules we wish to calculate the guest-guest interactions

molecules::Array{Molecule, 1}: An array of Molecule data structures

ljforcefield::LJForceField: A Lennard Jones forcefield data structure describing the interactions between different atoms

simulation_box::Box: The simulation box for the computation.

Returns

gg_energy::Float64: The guest-guest interaction energy of molecules[molecule_id] with the other molecules in molecules

PorousMaterials.vdw_energy_no_PBC — Function

potenergy = vdwenergynoPBC(atomsi, atomsj , ljff)

compute vdw potential energy without periodic boundary conditions